% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddBasisSCAN.R
\name{AddBasisSCAN}
\alias{AddBasisSCAN}
\title{AddBasisSCAN - add Fourier basis functions to SCAN dataset}
\usage{
AddBasisSCAN(df, num_basis)
}
\arguments{
\item{df}{a data frame for SCAN}

\item{num_basis}{an integer; the number of pairs of basis functions to add for year AND week}
}
\value{
a data frame; SCAN dataset with additional columns for basis function values
}
\description{
AddBasisSCAN - add Fourier basis functions to SCAN dataset
}
