% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddSewerSCAN.R
\name{AddSewerSCAN}
\alias{AddSewerSCAN}
\title{AddSewerSCAN - add sewer type information to SCAN dataset}
\usage{
AddSewerSCAN(df, sewer_df)
}
\arguments{
\item{df}{a data frame for SCAN}

\item{sewer_df}{a data frame of SCAN sewer type information}
}
\value{
a data frame; SCAN dataset with sewer type information cleaned and appended
}
\description{
Some changes are made manually:
- Oceanside Water Pollution Control Plant is combined
- Southeast San Francisco is \href{https://sfpuc.org/about-us/our-systems/sewer-system/our-combined-sewer}{combined}
- CODIGA is \href{https://suwater.stanford.edu/stormwater}{separate?}
- UC Davis is \href{https://sustainability.ucdavis.edu/goals/water}{separate?} (see also \href{https://facilities.ucdavis.edu/utilities-operations/campus-infrastructure-\%26-services#waste-water}{here})
}
