% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddWeatherSCAN.R
\name{AddWeatherSCAN}
\alias{AddWeatherSCAN}
\title{AddWeatherSCAN - add weather information to raw SCAN dataset}
\usage{
AddWeatherSCAN(df, weather_path = "data/NCDC")
}
\arguments{
\item{df}{a data frame for SCAN}

\item{weather_path}{a string; path to directory containing NOAA NCDC GHCN cleaned precipitation data by FIPS codes (no trailing '/')}
}
\value{
a data frame; subset and cleaned raw data from SCAN dataset
}
\description{
The following changes/edits are made:
- Add columns for average temperature/precipitation to scan data frame (average taken within day and over all counties served by that site)
- Add column for season (based upon meteorological season)
- Removes rows with NA in columns of interest (see below) to avoid imputation
}
