% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanNCDC.R
\name{CleanNCDC}
\alias{CleanNCDC}
\title{CleanNCDC - Clean .csv file from NOAA NCDC GHNC (yearly) direct download.}
\usage{
CleanNCDC(weather_path, fipscodes)
}
\arguments{
\item{weather_path}{a string; path to directory containing raw NOAA NCDC GHNC .csv files}

\item{fipscodes}{a vector of strings; list of fipscodes covered by SCAN dataset}
}
\value{
None
}
\description{
Cleans data sourced from \href{https://www.ncei.noaa.gov/pub/data/ghcn/daily/by_year/}{the NOAA NCDC}
daily climate data records by year. These were downloaded by hand.
Saves cleaned files as .csv with 'clean' prepended to file name.
}
