% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigureACF.R
\name{FigureACF}
\alias{FigureACF}
\title{FigureACF - Create and save ACF plot as PDF.}
\usage{
FigureACF(scan_fb, all_sites, data, out_path)
}
\arguments{
\item{scan_fb}{a data frame; cleaned SCAN data}

\item{all_sites}{a boolean flag; create and save ACF plots for individual sites}

\item{data}{a boolean flag; if ACF is for residuals or raw data}

\item{out_path}{a string; path to save figure}
}
\value{
None
}
\description{
Figure is a plot of the weighted average autocorrelations across all sites.
Function will also save a .csv file with the number of observations used to calculate the ACF at each lag.
}
