% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigureBoxplot.R
\name{FigureBoxplot}
\alias{FigureBoxplot}
\title{FigureBoxplot - Create and save boxplot of sites as PDF.}
\usage{
FigureBoxplot(scan_fb, out_path = "manu/figures")
}
\arguments{
\item{scan_fb}{a data frame; cleaned SCAN data}

\item{out_path}{a string; path to save figure}
}
\value{
None
}
\description{
Figure is a box and whisker plot of the SCAN data ordered along the x-axis by site longitude.
}
