% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigureComponents.R
\name{FigureComponents}
\alias{FigureComponents}
\title{FigureComponents - Create and save site components figure as PDF.}
\usage{
FigureComponents(
  scan_fb,
  site_name,
  start = "2022-01-01",
  end = "2023-04-01",
  limits = c(8.125, 8.44),
  model_path = "~/projects/pmmov-manu/stan/fits/wu1000_sep_lap_noucd/site_fits/",
  weather_path = "~/projects/pmmov-manu/data/NCDC",
  out_path = "manu/figures"
)
}
\arguments{
\item{scan_fb}{a data frame; cleaned SCAN data}

\item{site_name}{a string; name of site to plot}

\item{start}{a string; starting date of window for plotting}

\item{end}{a string; ending date of window for plotting}

\item{limits}{a vector; minimum and maximum values for y-axis}

\item{model_path}{a string; path where Stan model fit is saved}

\item{weather_path}{a string; path where cleaned weather data is saved}

\item{out_path}{a string; path to save figure}
}
\value{
None
}
\description{
Figure is a grid of the predicted time series made with Bayesian median model. Rug is colored by average daily precipitation for
the area.
}
