% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigurePeaks.R
\name{FigurePeaks}
\alias{FigurePeaks}
\title{FigurePeaks - Create and save peak plot as PDF.}
\usage{
FigurePeaks(scan_fb, peak_trough, lambda, out_path = "manu/figures")
}
\arguments{
\item{scan_fb}{a data frame; cleaned SCAN data}

\item{peak_trough}{a data frame; calculated peaks and troughs of sinusoids for each site}

\item{lambda}{a float; period length (in days)}

\item{out_path}{a string; path to save figure}
}
\value{
None
}
\description{
Figure is a scatterplot of peak dates for each site.
}
