% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigureQQPlot.R
\name{FigureQQPlot}
\alias{FigureQQPlot}
\title{FigureQQPlot - Create and save QQ plots as PDF.}
\usage{
FigureQQPlot(scan_fb, model, out_path = "manu/figures")
}
\arguments{
\item{scan_fb}{a data frame; cleaned SCAN data}

\item{model}{an \code{lm} object; fitted variance decomposition model}

\item{out_path}{a string; path to save figure}
}
\value{
None
}
\description{
Figure is a series of QQ plots of the variance decomposition model
standardized residual quantiles against the theoretical quantiles of the
standard Gaussian and standard Laplace.
}
