% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigureSewerComparison.R
\name{FigureSewerComparison}
\alias{FigureSewerComparison}
\title{FigureSewerComparison - Create and save sewer comparison figure as PDF.}
\usage{
FigureSewerComparison(
  scan_fb,
  sites = c("Calera Creek Water Recycling Plant",
    "Oceanside Water Pollution Control Plant"),
  start = "2022-11-01",
  end = "2023-03-01",
  weather_path = "~/pmmov-manu/data/NCDC",
  out_path = "manu/figures"
)
}
\arguments{
\item{scan_fb}{a data frame; cleaned SCAN data}

\item{sites}{a vector of strings; names of sites to plot}

\item{start}{a string; starting date of window for plotting}

\item{end}{a string; ending date of window for plotting}

\item{weather_path}{a string; path where cleaned weather data is saved}

\item{out_path}{a string; path to save figure}
}
\value{
None
}
\description{
Figure is raw observations from the specified sites with a LOESS smoother to exentuate
sewer system type differences.
}
