% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigureSitePred.R
\name{FigureSitePred}
\alias{FigureSitePred}
\title{FigureSitePred - Create and save site predictions as PDF.}
\usage{
FigureSitePred(
  scan_fb,
  pred_inters,
  prcp_inters,
  out_path = "manu/figures/site_preds"
)
}
\arguments{
\item{scan_fb}{a data frame; cleaned SCAN data}

\item{pred_inters}{a data frame; predictions from generalized Gaussian model (Model 4) for site(s)}

\item{prcp_inters}{a data frame; precipitation data for site(s)}

\item{out_path}{a string; path to save figure}
}
\value{
None
}
\description{
The site predictions are illustrated as a grid of the predicted time series made with Model 4.
Line is the median, bands are the 95\% prediction interval, and points are the observed values.
Rug is colored by average daily precipitation for the area.
}
