% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigureVariance.R
\name{FigureVariance}
\alias{FigureVariance}
\title{FigureVariance - Create and save variance decomposition figure as PDF.}
\usage{
FigureVariance(scan_fb, var_part, out_path = "manu/figures")
}
\arguments{
\item{scan_fb}{a data frame; cleaned SCAN data}

\item{var_part}{a data frame; variance partition for linear model (created with \code{\link{RunVarPart}})}

\item{out_path}{a string; path to save figure}
}
\value{
None
}
\description{
Figure is a stacked bar plot of the variance partition created from Model 3.
}
