% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakTrough.R
\name{PeakTrough}
\alias{PeakTrough}
\title{PeakTrough - Calculate peaks and troughs of Fourier bases for each site.}
\usage{
PeakTrough(
  scan_fb,
  lambda = 365.25,
  model_path = "~/projects/pmmov-manu/stan/wu1000_sep_lap_nucd/site_fits",
  out_path = "~/projects/pmmov-manu/checkpoints",
  plot_out_path = "~/projects/pmmov-manu/manu/figures/site_peak_trough"
)
}
\arguments{
\item{scan_fb}{a data frame; cleaned SCAN data}

\item{lambda}{a float; period for Fourier bases (non-negative)}

\item{model_path}{a string; path where individual Stan fits for each site are saved}

\item{out_path}{a string path; path to save output}

\item{plot_out_path}{a string path; path to save site plots}
}
\value{
None
}
\description{
PeakTrough - Calculate peaks and troughs of Fourier bases for each site.
}
