% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredInterval.R
\name{PredInterval}
\alias{PredInterval}
\title{PredInterval - Create predictions and bands for model fit with Stan.}
\usage{
PredInterval(scan_fb, beta, model_path, sites)
}
\arguments{
\item{scan_fb}{a data frame; cleaned WWTP data frame}

\item{beta}{a scalar; exponent for generalized Gaussian distribution}

\item{model_path}{a string path; path to RDS files for stan fits for each site}

\item{sites}{a vector of strings; names of sites to predict for}
}
\value{
a data frame of site metadata as well as quantiles and mean of predictions
using a random sample of coefficient estimates across all chains from Stan fit
}
\description{
PredInterval - Create predictions and bands for model fit with Stan.
}
