% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepData.R
\name{PrepData}
\alias{PrepData}
\title{PrepData - Clean and arrange data frame for PMMoV manuscript.}
\usage{
PrepData(
  out_path,
  raw_path = "data/scan",
  weather_path = "data/NCDC",
  clean_ncdc = FALSE
)
}
\arguments{
\item{out_path}{a string; path to save cleaned data .csv file}

\item{raw_path}{a string; path to raw WastewaterSCAN data .csv}

\item{weather_path}{a string; path to saved cleaned weather data .csv files}

\item{clean_ncdc}{a boolean flag; whether to aggregate the NCDC weather data}
}
\value{
a data frame; the cleaned SCAN data used for the analyses
}
\description{
PrepData - Clean and arrange data frame for PMMoV manuscript.
}
