% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocRawSCAN.R
\name{PreprocRawSCAN}
\alias{PreprocRawSCAN}
\title{PreprocRawSCAN - preprocessing of raw SCAN dataset}
\usage{
PreprocRawSCAN(df)
}
\arguments{
\item{df}{a data frame for SCAN}
}
\value{
a data frame; subset and cleaned raw data from SCAN dataset
}
\description{
The following changes/edits are made:
- Several entries are missing zip/FIPS codes. These are added manually.
- Removes sample 'COD110099 influent'
- Collection_Date is cast as 'Date' type
- County_FIPS is renamed as fipscode
- Collection_Date is renamed as date
- missing FIPS codes are changed to 'NA'
- select relevant columns (sampling location/time metadata, SC2 N gene, PMMoV)
- add latitude and longitude of sites by zipcode
- filters data to after change in \href{https://docs.google.com/document/d/1qiNq3wh0H8GrELesgLUrDZF9oTdanI40CZ6tEZSUbic/edit}{lab methods}
- add column for log10 PMMoV
}
