% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepStanData.R
\name{PrepStanData}
\alias{PrepStanData}
\title{PrepStanData - Arrange data frame for Stan sampling.}
\usage{
PrepStanData(site_data, weather_path = "data/NCDC")
}
\arguments{
\item{site_data}{a data frame; site-specific SCAN data}

\item{weather_path}{a string; path to saved cleaned weather data .csv files}
}
\value{
a data frame; the prepared SCAN data used for the Stan sampling of response
}
\description{
PrepStanData - Arrange data frame for Stan sampling.
}
