% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDistMod.R
\name{RunDistMod}
\alias{RunDistMod}
\title{RunDistMod - Fit and save quantile regression model with distance covariates.}
\usage{
RunDistMod(out_path, scan_fb)
}
\arguments{
\item{out_path}{a string; path to save fitted models}

\item{scan_fb}{a data frame; cleaned WWTP data frame}
}
\value{
None
}
\description{
RunDistMod - Fit and save quantile regression model with distance covariates.
}
