% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunIDW.R
\name{RunIDW}
\alias{RunIDW}
\title{RunIDW - Do IDW spatial interpolation.}
\usage{
RunIDW(out_path, scan_fb, pred_grid, power = NA, use_last = TRUE)
}
\arguments{
\item{out_path}{a string; path to save fitted models}

\item{scan_fb}{a data frame; cleaned WWTP data frame}

\item{pred_grid}{a data frame; coordinates at which to interpolate (should contain columns for latitude and longitude)}

\item{power}{a positive scalar; power to use for IDW}

\item{use_last}{a boolean flag; if TRUE, use the last saved interpolation}
}
\value{
a data frame of interpolated values at the locations specified by \code{pred_grid}
}
\description{
Code is a (very slightly) modified version of the code from the \code{phylin} package.
}
\details{
Tarroso, P., Velo-Anton, G., & Carvalho, S. B. (2015). PHYLIN: an R
package for phylogeographic interpolation. Molecular Ecology Resources, 15(2), 349-357.

Tarroso, P., Carvalho, S. B. & Velo-Anton, G. (2019). PHYLIN 2.0:
Extending the phylogeographic interpolation method to include
uncertainty and user-defined distance metrics. Molecular Ecology Resources, 19(2), 1081-1094.

There is an error when calculating the distance matrix where the matrix is cast to a list
which throws an indexing error.

The power parameter will be chosen by LOOCV if no value is provided.
}
