% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunPred.R
\name{RunPred}
\alias{RunPred}
\title{RunPred - Create predictions and bands for model fit with Stan.}
\usage{
RunPred(
  out_path,
  weather_path,
  model_path,
  scan_fb,
  beta,
  sites,
  prefix = "cross_us"
)
}
\arguments{
\item{out_path}{a string path; path to save output}

\item{weather_path}{a string path; path to cleaned weather data .csv files from NCDC}

\item{model_path}{a string path; path to RDS files for stan fits for each site}

\item{scan_fb}{a data frame; cleaned WWTP data frame}

\item{beta}{a scalar; exponent for generalized Gaussian distribution}

\item{sites}{a vector of strings; names of sites to predict for}

\item{prefix}{a string; prefix for saved RDS files}
}
\value{
None
}
\description{
RunPred - Create predictions and bands for model fit with Stan.
}
