% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunQuantReg.R
\name{RunQuantReg}
\alias{RunQuantReg}
\title{RunQuantReg - Fit and save linear quantile regression models.}
\usage{
RunQuantReg(out_path, scan_fb)
}
\arguments{
\item{out_path}{a string; path to save fitted models}

\item{scan_fb}{a data frame; cleaned WWTP data frame}
}
\value{
None
}
\description{
RunQuantReg - Fit and save linear quantile regression models.
}
