% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunVarPart.R
\name{RunVarPart}
\alias{RunVarPart}
\title{RunVarPart - Calculate variance partitions for linear model.}
\usage{
RunVarPart(scan_fb, out_path)
}
\arguments{
\item{scan_fb}{a data frame; cleaned SCAN data}

\item{out_path}{a string; path to save .csv of coefficients}
}
\value{
None
}
\description{
RunVarPart - Calculate variance partitions for linear model.
}
