% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveCoefs.R
\name{SaveCoefs}
\alias{SaveCoefs}
\title{SaveCoefs - Save model coefficients as .csv}
\usage{
SaveCoefs(
  model,
  out_path,
  file_name,
  index_fix = c("lat", "lng", "avg_prcp", "sewer", "avg_prcp:sewer", "sin1_wk",
    "cos1_wk", "sin2_wk", "cos2_wk")
)
}
\arguments{
\item{model}{a model}

\item{out_path}{a string; path to save .csv of coefficients}

\item{file_name}{a string; name of .csv file (no extension)}

\item{index_fix}{a vector; contains strings for names of fixed effect variables}
}
\value{
None
}
\description{
SaveCoefs - Save model coefficients as .csv
}
