% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveStanCoefs.R
\name{SaveStanCoefs}
\alias{SaveStanCoefs}
\title{SaveStanCoefs - save summaries of model parameters for all sites}
\usage{
SaveStanCoefs(scan_fb, model_path, out_path)
}
\arguments{
\item{scan_fb}{a data frame; cleaned WWTP data framed}

\item{model_path}{a string; path to saved Stan model fits}

\item{out_path}{a string; path to save .csv file}
}
\value{
None
}
\description{
SaveStanCoefs - save summaries of model parameters for all sites
}
