% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveSummary.R
\name{SaveSummary}
\alias{SaveSummary}
\title{SaveSummary - save summary statistics for SCAN dataset to .csv file.}
\usage{
SaveSummary(out_path, scan_fb)
}
\arguments{
\item{out_path}{a string; path to save .csv file}

\item{scan_fb}{a data frame; cleaned WWTP data frame}
}
\value{
None
}
\description{
SaveSummary - save summary statistics for SCAN dataset to .csv file.
}
