% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteAbbrev.R
\name{SiteAbbrev}
\alias{SiteAbbrev}
\title{SiteAbbrev}
\usage{
SiteAbbrev(scan_fb, out_path = "checkpoints/csv")
}
\arguments{
\item{scan_fb}{a data frame; cleaned SCAN data}

\item{out_path}{a string; path to save figure}
}
\value{
site_table; a data frame with site name abbreviations
}
\description{
Make site abbreviation table
}
