% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteResid.R
\name{SiteResid}
\alias{SiteResid}
\title{SiteResid - Calculate residuals from Stan models.}
\usage{
SiteResid(scan_fb, model_path, out_path = "~/pmmov-manu/checkpoints")
}
\arguments{
\item{scan_fb}{a data frame; cleaned WWTP data frame}

\item{model_path}{a string path; path to RDS files for stan fits for each site}

\item{out_path}{a string path; path to save data frame of residuals to}
}
\value{
none
}
\description{
SiteResid - Calculate residuals from Stan models.
}
