% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigureGradient.R
\name{FigureGradient}
\alias{FigureGradient}
\title{FigureGradient - Create and save gradient map figure as PDF.}
\usage{
FigureGradient(scan_fb, model, out_path = "manu/figures")
}
\arguments{
\item{scan_fb}{a data frame; cleaned SCAN data}

\item{model}{an \code{rq} object; fitted simple quantile regression model (Model 1)}

\item{out_path}{a string; path to save figure}
}
\value{
None
}
\description{
Figure is a map of the contiguous USA with a color gradient made according
to the predicted log PMMoV made with the simple quantile regression model (Model 1).
}
